Server/requirements.txt
dify_client
azure-cognitiveservices-speech
python-dotenv
paho-mqtt